requirements-harmony.txt
# Harmony Agent Dependencies
httpx>=0.27,<1.0          # Async HTTP client for API calls
pydantic>=2.11,<3.0       # Data validation and models
python-dotenv>=1.1,<2.0   # Configuration management
rapidfuzz>=3.14,<4.0      # Fuzzy string matching for author/series detection
tenacity>=9.1,<10.0       # Retry logic with exponential backoff
tqdm>=4.67,<5.0           # Progress bars

# Testing
pytest>=8.4,<9.0
pytest-asyncio>=1.2,<2.0
respx>=0.22,<1.0          # HTTP mocking for tests
